fix(index-template): use int64 for Priority/Version, restore metadata test
models.IndexTemplate.Priority and Version were *int, causing potential
truncation on 32-bit builds and a type-width mismatch with the typed
go-elasticsearch client's *int64 fields. Widen to *int64 and remove the
int() conversion in expand.go.

Restore the metadata round-trip assertion in the flatten test that was
dropped when migrating from models.IndexTemplate to estypes.IndexTemplate.

Revert apostrophe-only whitespace change to docs/resources/elasticsearch_data_stream.md
which was unassociated with this change.

+++ internal/models/models.go
@@ -173,9 +173,9 @@ type IndexTemplate struct {
 	DataStream                      *DataStreamSettings `json:"data_stream,omitempty"`
 	IndexPatterns                   []string            `json:"index_patterns"`
 	Meta                            map[string]any      `json:"_meta,omitempty"`
-	Priority                        *int                `json:"priority,omitempty"`
+	Priority                        *int64              `json:"priority,omitempty"`
 	Template                        *Template           `json:"template,omitempty"`
-	Version                         *int                `json:"version,omitempty"`
+	Version                         *int64              `json:"version,omitempty"`
 }
 
 type DataStreamSettings struct {
